[repo] Split tracker legacy link plugin to lib
